Finalizando controle de usuario

+++ resources/views/painel/user/index.blade.php
@@ -54,9 +54,9 @@
                             <td>{{$user->tipo}}</td>
                             <td>{{$user->status}}</td>
                             <td width="150px">
-                                <a href="/painel/user/detail/{{$user->id}}" class="btn btn-default" alt="Exibir o usuário" title="Exibir o usuário"><span class="glyphicon glyphicon-eye-open" aria-hidden="true"></span></a>
-                                <a href="/painel/user/edit/{{$user->id}}" class="btn btn-info" alt="Editar o usuário" title="Editar o usuário"><span class="glyphicon glyphicon-pencil" aria-hidden="true"></span></a>
-                                <a href="/painel/user/delete/{{$user->id}}" class="btn btn-danger" alt="Excluir o usuário" title="Excluir o usuário"><span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a>
+                                <a href="/painel/user/detail/{{$user->id}}" class="btn btn-success" alt="Exibir o usuário" title="Exibir o usuário"><span class="glyphicon glyphicon-eye-open" aria-hidden="true"></span></a>
+                                <a href="/painel/user/edit/{{$user->id}}" class="btn btn-warning" alt="Editar o usuário" title="Editar o usuário"><span class="glyphicon glyphicon-pencil" aria-hidden="true"></span></a>
+                                <a href="/painel/user/delete/{{$user->id}}" onclick="return confirm('Realmente deseja excluir este usuário?')" class="btn btn-danger" alt="Excluir o usuário" title="Excluir o usuário"><span class="glyphicon glyphicon-remove" aria-hidden="true"></span></a>
                             </td>
                         </tr>
                     @empty
